refactor(output): serialize integral metrics as integers
Integer-valued metrics now serialize as integers instead of floats, and
their public `Stats` accessors return `u64` instead of `f64`. Affected:
every count/sum/min/max (cyclomatic, cognitive, exit, nargs, nom, tokens,
loc lines, ABC assignments/branches/conditions, npa/npm counts), Halstead
`length`/`vocabulary` and the four operator/operand counts, and all WMC
values. Ratios, averages, ABC magnitude, the derived Halstead scores, and
MI remain `f64`.

JSON/TOML/YAML now emit `"sloc": 5` rather than `5.0`, CBOR encodes these
as compact integers rather than float64, and CSV is unchanged (it already
stripped the trailing `.0`). Float-consuming callers (averages, MI, WMC
aggregation, Halstead derived formulas, the CSV row, dump, thresholds)
cast back with `as f64`, so no metric value changes — only its type.

This is a SemVer-breaking shape change to the serialized output and the
accessor signatures; the CHANGELOG entry is marked (breaking) and deferred
to the 2.0.0 release.

Fixes #530

+++ big-code-analysis-cli/src/markdown_report.rs
@@ -159,8 +159,8 @@ fn extract_summaries_inner(
             lloc: m.loc.lloc() as usize,
             cloc: m.loc.cloc() as usize,
             tokens: m.tokens.tokens_sum() as usize,
-            cyclomatic: m.cyclomatic.cyclomatic(),
-            cognitive: m.cognitive.cognitive(),
+            cyclomatic: m.cyclomatic.cyclomatic() as f64,
+            cognitive: m.cognitive.cognitive() as f64,
             halstead_volume: m.halstead.volume(),
             halstead_difficulty: m.halstead.difficulty(),
             halstead_effort: m.halstead.effort(),
@@ -173,9 +173,9 @@ fn extract_summaries_inner(
             nexits: m.nexits.exit_sum() as usize,
             nom: m.nom.total() as usize,
             abc: m.abc.magnitude(),
-            wmc: m.wmc.total_wmc(),
-            npa: m.npa.total_npa(),
-            npm: m.npm.total_npm(),
+            wmc: m.wmc.total_wmc() as f64,
+            npa: m.npa.total_npa() as f64,
+            npm: m.npm.total_npm() as f64,
         });
 
         stack.extend(current.spaces.iter().rev());
